+++ Network/NetworkBuffer.hh
@@ -19,14 +19,14 @@ public:
   }
 
   virtual IBuffer	&operator<<(const int &);
-  virtual IBuffer	&operator<<(const unsigned int &);
+  virtual IBuffer	&operator<<(const unsigned long &);
   virtual IBuffer	&operator<<(const float &);
   virtual IBuffer	&operator<<(const std::string &);
   virtual IBuffer	&operator<<(const std::size_t &);
   virtual IBuffer	&operator<<(const char &);
 
   virtual IBuffer	&operator>>(int &);
-  virtual IBuffer	&operator>>(unsigned int &);
+  virtual IBuffer	&operator>>(unsigned long &);
   virtual IBuffer	&operator>>(float &);
   virtual IBuffer	&operator>>(std::string &);
   virtual IBuffer	&operator>>(std::size_t &);
